python/dao-voting-dashboard/pyproject.toml
[project]
name = "dao-voting-dashboard"
version = "0.1.0"
description = "Off-chain + on-chain DAO voting dashboard for BNB Chain"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "web3>=7.0.0",
    "eth-account>=0.13.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
